feat: 다양한 모양 대응을 위해 Brush 인터페이스화
